+++ example/lib/saved_image_view.dart
@@ -11,13 +11,12 @@ class SavedImageView extends StatefulWidget {
 }
 
 class _SavedImageViewState extends State<SavedImageView> {
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-          backgroundColor: const Color.fromARGB(255, 25, 40, 46),
-          title: const Text("WhatsApp Story Editor Results")),
-      backgroundColor: Colors.black,
+          title: const Text("Story Creator Result")),
       body: Padding(
         padding: const EdgeInsets.symmetric(horizontal: 10),
         child: Column(