+++ lib/src/views/main_view.dart
@@ -26,33 +26,63 @@ class _MainControllerViewState extends State<MainControllerView> {
 
   @override
   Widget build(BuildContext context) {
-    //moved Material App from here to whatsappstoryeditor screen
-    return SafeArea(
-      child: Scaffold(
-          backgroundColor: Theme.of(context).colorScheme.background,
-          body: Obx(
-            () => Column(
-                    children: [
-                      Expanded(
-                        child: Stack(
-                          children: [
-                                Center(
-                                    child: BackgroundImage(
-                                        context: context, file: widget.file)),
+    return WillPopScope(
+      onWillPop: () async {
+        await showDialog(
+          context: context,
+          builder: (context) {
+            return AlertDialog(
+              title: const Text("Exit"),
+              content: const Text("Are you sure you want to exit?"),
+              actions: [
+                TextButton(
+                  onPressed: () {
+                    Navigator.pop(context);
+                  },
+                  child: const Text("No"),
+                ),
+                TextButton(
+                  onPressed: () {
+                    Get.deleteAll();
+                    Navigator.pop(context);
+                    Navigator.pop(context);
+                    Navigator.pop(context);
+                  },
+                  child: const Text("Yes"),
+                ),
+              ],
+            );
+          },
+        );
+        return false;
+      },
+      child: SafeArea(
+        child: Scaffold(
+            backgroundColor: Theme.of(context).colorScheme.background,
+            body: Obx(
+              () => Column(
+                      children: [
+                        Expanded(
+                          child: Stack(
+                            children: [
+                                  Center(
+                                      child: BackgroundImage(
+                                          context: context, file: widget.file)),
 
-                            controller.editingModeSelected ==
-                                    EDITINGMODE.DRAWING
-                                ? const PaintView(shouldShowControls: true)
-                                : controller.editingModeSelected ==
-                                        EDITINGMODE.TEXT
-                                    ? const TextView()
-                                    : BasicView(file: widget.file)
-                          ],
-                        ),
-                      )
-                    ],
-                  ),
-          )),
+                              controller.editingModeSelected ==
+                                      EDITINGMODE.DRAWING
+                                  ? const PaintView(shouldShowControls: true)
+                                  : controller.editingModeSelected ==
+                                          EDITINGMODE.TEXT
+                                      ? const TextView()
+                                      : BasicView(file: widget.file)
+                            ],
+                          ),
+                        )
+                      ],
+                    ),
+            )),
+      ),
     );
   }
 }